+++ src/pages/DrinkDetails.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { useParams, useHistory, useRouteMatch } from 'react-router-dom';
 import copy from 'clipboard-copy';
+import DrinkPreparation from '../components/DrinkPreparation';
 import fetchRecipeByDetails from '../RequisiçõesAPI/drink/RequestByDetails';
 import shareIcon from '../images/shareIcon.svg';
 import whiteHeartIcon from '../images/whiteHeartIcon.svg';
@@ -9,7 +10,7 @@ import FoodsRecomendation from '../components/FoodsRecomendation';
 // https://dev.to/marcelomatosdev/react-adding-a-video-player-to-play-youtube-videos-in-your-project-30p
 
 export default function DrinkDetails() {
-  const fifteen = 15;
+  // const fifteen = 15;
   const { url } = useRouteMatch();
   const bottomFixed = {
     position: 'fixed',
@@ -24,12 +25,12 @@ export default function DrinkDetails() {
   const recipeId = pathname.split('/')[2];
 
   useEffect(() => {
-    const handleSelectedFood = async () => {
+    const handleSelectedDrink = async () => {
       const response = await fetchRecipeByDetails(recipeId);
       const result = await response.drinks;
       setRecipeDetails(result[0]);
     };
-    handleSelectedFood();
+    handleSelectedDrink();
   }, []);
 
   useEffect(() => {
@@ -44,22 +45,24 @@ export default function DrinkDetails() {
     setIsCopy(true);
   };
 
-  const preparation = () => {
-    const ingredientsList = [];
-    for (let index = 1; index <= fifteen; index += 1) {
-      if (recipeDetails[`strIngredient${index}`] !== ''
-      && recipeDetails[`strIngredient${index}`] !== null
-      ) {
-        ingredientsList.push(
-          `${recipeDetails[`strIngredient${index}`]}: 
-            ${recipeDetails[`strMeasure${index}`]}`,
-        );
-      }
-    }
-    return ingredientsList;
-  };
+  // const preparation = () => {
+  //   const ingredientsList = [];
+  //   for (let index = 1; index <= fifteen; index += 1) {
+  //     if (recipeDetails[`strIngredient${index}`] !== ''
+  //     && recipeDetails[`strIngredient${index}`] !== null
+  //     ) {
+  //       ingredientsList.push(
+  //         `${recipeDetails[`strIngredient${index}`]}:
+  //           ${recipeDetails[`strMeasure${index}`]}`,
+  //       );
+  //     }
+  //   }
+  //   return ingredientsList;
+  // };
 
-  const handleFavorite = ({ idDrink, strAlcoholic, strCategory, strDrink, strDrinkThumb }) => {
+  const handleFavorite = (
+    { idDrink, strAlcoholic, strCategory, strDrink, strDrinkThumb },
+  ) => {
     if (!isFavorite) {
       localStorage.setItem('favoriteRecipes', JSON.stringify([
         {
@@ -117,14 +120,15 @@ export default function DrinkDetails() {
       </button>
       <p data-testid="recipe-category">{ recipeDetails.strAlcoholic }</p>
       <ul>
-        { preparation().map((ingredient, index) => (
+        <DrinkPreparation recipeId={ recipeId } />
+        {/* { preparation().map((ingredient, index) => (
           <li
             data-testid={ `${index}-ingredient-name-and-measure` }
             key={ index }
           >
             {ingredient}
           </li>
-        ))}
+        ))} */}
       </ul>
       <p data-testid="instructions">{recipeDetails.strInstructions}</p>
       <FoodsRecomendation />
